requirements.txt
google-generativeai
requests
python-dotenv
Pillow
gpiozero
